+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - This CHANGELOG
 
+### Changed
+
+- `table()` in QueryBuilder will throw an exception
+
 ## [1.0.0-rc.1] - 2020-05-11
 
 ### Changed